Rename gen_ts to generic_time_ops
