+++ raphtory/src/db/api/storage/graph/storage_ops/time_semantics.rs
@@ -14,7 +14,7 @@ use rayon::iter::ParallelIterator;
 use std::ops::Range;
 use storage::{
     api::graph_props::{GraphPropEntryOps, GraphPropRefOps},
-    gen_ts::ALL_LAYERS,
+    generic_time_ops::ALL_LAYERS,
 };
 
 impl GraphTimeSemanticsOps for GraphStorage {
